+++ nifi-processors/nifi-health-patterns-bundle/nifi-health-patterns-processors/src/test/java/com/ibm/healthpatterns/processors/deid/DeIdentifyAndPostToFHIRTest.java
@@ -22,18 +22,17 @@ import static org.junit.Assert.assertTrue;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.nifi.util.MockFlowFile;
-import org.apache.nifi.util.TestRunner;
 import org.apache.nifi.util.TestRunners;
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.ibm.healthpatterns.processors.common.FHIRCustomProcessorTest;
+import com.ibm.healthpatterns.processors.common.FHIRServiceCustomProcessor;
 
 /**
  * This test runs the custom process against a real de-id service and FHIR server dedicated 
@@ -44,29 +43,17 @@ import com.fasterxml.jackson.databind.ObjectMapper;
  * 
  * @author Luis A. García
  */
-public class DeIdentifyAndPostToFHIRTest {
-
-	private TestRunner testRunner;
+public class DeIdentifyAndPostToFHIRTest extends FHIRCustomProcessorTest {
 
 	private String deidURL;
-	private String fhirURL;
-	private String fhirUsername;
-	private String fhirPassword;
-
-	private List<String> createdResources;
-
 	private DeIdentifyAndPostToFHIR customProcessor;
-
+	
 	/**
 	 * 
 	 */
 	public DeIdentifyAndPostToFHIRTest() {
-		// These URLs are for two IBM Cloud services that we setup specifically to run these tests 
+		// This URL is for IBM Cloud services that we setup specifically to run these tests 
 		deidURL = "http://3a5d0fa4-us-south.lb.appdomain.cloud:8080/api/v1";
-		fhirURL = "http://4603f72b-us-south.lb.appdomain.cloud/fhir-server/api/v4";
-		fhirUsername = "fhiruser";
-		fhirPassword = "integrati0n";
-		createdResources = new ArrayList<String>();
 	}
 
 	/**
@@ -77,27 +64,6 @@ public class DeIdentifyAndPostToFHIRTest {
 	public void init() throws IOException {
 		customProcessor = new DeIdentifyAndPostToFHIR();
 		testRunner = TestRunners.newTestRunner(customProcessor);
-	}
-
-	/**
-	 * @throws IOException 
-	 * 
-	 */
-	@After
-	public void cleanUpFHIR() throws IOException {
-		for (String uri : createdResources) {
-			String[] uriElements = uri.split("/");
-			String id = uriElements[uriElements.length - 3];
-			String type = uriElements[uriElements.length - 4];
-			System.out.println("Deleting " + type + " resource " + id);
-			customProcessor.getDeidentifier()
-			.getFhirClient()
-			.delete()
-			.resourceById(type, id)
-			.execute();
-		}
-		System.out.println("------------------");
-		System.out.println();    	
 	}
 
 	/**
@@ -204,4 +170,11 @@ public class DeIdentifyAndPostToFHIRTest {
 		testRunner.assertNotValid();
 	}
 
+	/* (non-Javadoc)
+	 * @see com.ibm.healthpatterns.processors.deid.FHIRCustomProcessorTest#getCustomProcessor()
+	 */
+	@Override
+	public FHIRServiceCustomProcessor getCustomProcessor() {
+		return customProcessor;
+	}
 }
