+++ services/common/.gitignore
@@ -1,0 +1,16 @@
+/target
+/.m2
+
+
+/.classpath
+/.project
+/.settings
+
+/caches
+/local.properties
+.*.swp
+.DS_Store
+/src/main/resources/application-local.properties
+
+src/main/resources/application-local.properties
+src/main/resources/localdev-config.json
