+++ services/common/README.md
@@ -1,0 +1,29 @@
+<p align="center">
+    <a href="https://github.com/Alvearie">
+        <img src="https://avatars.githubusercontent.com/u/72946463?s=200&v=4" height="100" alt="Alvearie">
+    </a>
+</p>
+
+
+<p align="center">
+    <a href="https://www.ibm.com/developerworks/learn/java/">
+    <img src="https://img.shields.io/badge/platform-java-lightgrey.svg?style=flat" alt="platform">
+    </a>
+    <img src="https://img.shields.io/badge/license-Apache2-blue.svg?style=flat" alt="Apache 2">
+</p>
+
+
+# Common Component
+
+Some common classes used by the other Health Pattern Services.
+
+
+## Building
+
+See instructions in parent.
+
+## License
+
+This application is licensed under the Apache License, Version 2. Separate third-party code objects invoked within this code pattern are licensed by their respective providers pursuant to their own separate licenses. Contributions are subject to the [Developer Certificate of Origin, Version 1.1](https://developercertificate.org/) and the [Apache License, Version 2](https://www.apache.org/licenses/LICENSE-2.0.txt).
+
+[Apache License FAQ](https://www.apache.org/foundation/license-faq.html#WhatDoesItMEAN)
