+++ services/common/pom.xml
@@ -23,21 +23,23 @@
         <version>0.0.1-SNAPSHOT</version>
     </parent>
 
-    <artifactId>terminology-service-core</artifactId>
-    <name>FHIR Terminology Service Core</name>
+    <artifactId>common</artifactId>
+    <name>Health Patterns Common Component</name>
 
-    <dependencies>
-        <dependency>
-            <groupId>com.ibm.healthpatterns</groupId>
-            <artifactId>common</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>com.ibm.healthpatterns</groupId>
-            <artifactId>common</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-            <type>test-jar</type>
-            <scope>test</scope>
-        </dependency>        
-    </dependencies>    
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-jar-plugin</artifactId>
+				<version>2.2</version>
+				<executions>
+					<execution>
+						<goals>
+							<goal>test-jar</goal>
+						</goals>
+					</execution>
+				</executions>
+			</plugin>
+		</plugins>
+	</build>
 </project>
