+++ services/deid-core/pom.xml
@@ -26,11 +26,18 @@
     <artifactId>deid-core</artifactId>
     <name>De-Identification Core</name>
 
-	<dependencies>
-	   <dependency>
-		  <groupId>com.fasterxml.jackson.core</groupId>
-		  <artifactId>jackson-core</artifactId>
-		  <version>2.11.2</version>
-	   </dependency>
-	</dependencies>
+    <dependencies>
+        <dependency>
+	        <groupId>com.ibm.healthpatterns</groupId>
+	        <artifactId>common</artifactId>
+	        <version>0.0.1-SNAPSHOT</version>
+        </dependency>
+		<dependency>
+            <groupId>com.ibm.healthpatterns</groupId>
+            <artifactId>common</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+			<type>test-jar</type>
+			<scope>test</scope>
+		</dependency>
+    </dependencies>
 </project>
